% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{d2}
\alias{d2}
\title{second dataset for HW 3}
\format{An object of class \code{data.frame} with 142 rows and 6 columns.}
\usage{
d2
}
\description{
second dataset for HW 3
}
\references{
\url{http://users.stat.umn.edu/~almquist/3811_examples/gapminder2007ex.csv}
}
\author{
Zack W. Almquist \email{almquist@umn.edu}
}
\keyword{datasets}
