% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{d3}
\alias{d3}
\title{Third dataset for HW 3}
\format{An object of class \code{data.frame} with 5 rows and 1 columns.}
\usage{
d3
}
\description{
Third dataset for HW 3
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p1.txt}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
