% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuyaoLiuToolsmyplot.R
\name{dplyrWrapper}
\alias{dplyrWrapper}
\title{dplyr Wrapper}
\usage{
dplyrWrapper(x)
}
\arguments{
\item{x}{A data-frame}
}
\value{
data-frame
}
\description{
This is a wrapper for DPLYR
}
\examples{
data(d)
dplyrWrapper(d)
}
