% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuyaoLiuToolsRfunctions.R
\name{func8}
\alias{func8}
\title{$x^T A^{-1} x$ and GIEMO}
\usage{
func8(a, x)
}
\arguments{
\item{a}{numeric matrix}

\item{x}{vector}
}
\value{
scalar
}
\description{
Given a numeric matrix A and a numeric vector `x`, 
calculates $x^T A^{-1} x$
Note that this only makes sense when A is a square matrix 
and the dimension of x is the same as 
the dimensions of the row and column dimensions of A
}
\details{
Follow GIEMO
}
\examples{
x <- c(1,2,3,4)
a <-matrix(rnorm(16),nc=4,nr=4)

result8 <- func8(a, x)
result8

}
